Testing on WebMvc complete

+++ src/main/java/ru/hogwarts/school/model/Faculty.java
@@ -12,7 +12,7 @@ public class Faculty {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "id", nullable = false)
-    private int id;
+    private Long id;
     private String name;
     private String color;
 
@@ -21,7 +21,7 @@ public class Faculty {
 
 
     public Faculty(int id,String name, String color) {
-        this.id = id;
+        this.id = (long) id;
         this.name = name;
         this.color = color;
     }
@@ -33,11 +33,11 @@ public class Faculty {
     public Faculty(long l, String anything, String anyColor) {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
